requirements.txt
gym==0.12.5
pyyaml==5.1.1
pytest==4.6.3